--- positive-health-int-team-2-frontend/src/components/Geolocation.svelte
@@ -1,4 +1,0 @@
-<script>
-
-
-</script>
